move index trait
