+++ rust/lance-index/Cargo.toml
@@ -20,6 +20,8 @@ arrow-arith.workspace = true
 arrow-select.workspace = true
 async-trait.workspace = true
 datafusion-common.workspace = true
+datafusion-expr = "32.0.0"
+datafusion-physical-expr = { version = "32.0.0", default-features = false }
 futures.workspace = true
 half.workspace = true
 lance-arrow.workspace = true
@@ -28,14 +30,14 @@ lance-linalg.workspace = true
 log.workspace = true
 num_cpus.workspace = true
 num-traits.workspace = true
+object_store.workspace = true
 prost.workspace = true
 rand.workspace = true
+serde_json.workspace = true
+serde.workspace = true
 snafu.workspace = true
 tokio.workspace = true
 tracing.workspace = true
-datafusion-physical-expr = { version = "32.0.0", default-features = false }
-datafusion-expr = "32.0.0"
-object_store.workspace = true
 
 [dev-dependencies]
 approx.workspace = true
